+++ packages/sdk/src/config/e2e.ts
@@ -1,6 +1,7 @@
 import { toAddress, ZERO_ADDRESS } from "@rarible/types"
 import { id32 } from "../common/id"
 import type { EthereumConfig } from "./type"
+import { FEE_CONFIG_URL } from "./common"
 
 export const e2eConfig: EthereumConfig = {
 	basePath: "https://ethereum-api-e2e.rarible.org",
@@ -18,9 +19,7 @@ export const e2eConfig: EthereumConfig = {
 		openseaV1: ZERO_ADDRESS,
 		cryptoPunks: toAddress("0x3268d39B77A7F189209E87E13eedA613098b32c7"),
 	},
-	fees: {
-		v2: 0,
-	},
+	feeConfigUrl: FEE_CONFIG_URL,
 	openSea: {
 		metadata: id32("RARIBLE"),
 		proxyRegistry: ZERO_ADDRESS,
@@ -30,4 +29,5 @@ export const e2eConfig: EthereumConfig = {
 		erc1155: toAddress("0xF35BC450544C7d8c2559B4DAeb51E9617F39e7C6"),
 	},
 	weth: toAddress("0xc6f33b62a94939e52e1b074c4ac1a801b869fdb2"),
+	auction: toAddress("0xE091e5511815107072ed1C42d0d00f151D6197ed"),
 }