+++ packages/sdk/src/nft/burn.test.ts
@@ -9,7 +9,7 @@ import {
 import { toBn } from "@rarible/utils"
 import type { Ethereum } from "@rarible/ethereum-provider"
 import { checkAssetType as checkAssetTypeTemplate } from "../order/check-asset-type"
-import { send as sendTemplate } from "../common/send-transaction"
+import { getSendWithInjects } from "../common/send-transaction"
 import { getApiConfig } from "../config/api-config"
 import { createTestProviders } from "../common/create-test-providers"
 import {
@@ -21,8 +21,9 @@ import {
 import { createEthereumApis } from "../common/apis"
 import { getEthereumConfig } from "../config"
 import { checkChainId as checkChainIdTemplate } from "../order/check-chain-id"
-import type { ERC1155RequestV1, ERC721RequestV2, ERC721RequestV3, ERC1155RequestV2} from "./mint"
-import { mint as mintTemplate } from "./mint"
+import { retry } from "../common/retry"
+import type { ERC1155RequestV1, ERC1155RequestV2, ERC721RequestV2, ERC721RequestV3 } from "./mint"
+import { mint as mintTemplate, MintResponseTypeEnum } from "./mint"
 import { signNft } from "./sign-nft"
 import { burn as burnTemplate } from "./burn"
 import { ERC1155VersionEnum, ERC721VersionEnum } from "./contracts/domain"
@@ -40,10 +41,10 @@ describe.each(providers)("burn nfts", (ethereum: Ethereum) => {
 	const mintLazyApi = new NftLazyMintControllerApi(configuration)
 	const gatewayApi = new GatewayControllerApi(configuration)
 	const sign = signNft.bind(null, ethereum, 17)
-	const send = sendTemplate.bind(ethereum, gatewayApi)
-	const checkAssetType = checkAssetTypeTemplate.bind(null, collectionApi)
 	const config = getEthereumConfig("e2e")
 	const checkChainId = checkChainIdTemplate.bind(null, ethereum, config)
+	const send = getSendWithInjects().bind(ethereum, gatewayApi, checkChainId)
+	const checkAssetType = checkAssetTypeTemplate.bind(null, collectionApi)
 	const mint = mintTemplate.bind(null, ethereum, send, sign, collectionApi)
 	const burn = burnTemplate.bind(null, ethereum, send, checkAssetType, apis)
 	const contractErc721 = toAddress("0x87ECcc03BaBC550c919Ad61187Ab597E9E7f7C21")
@@ -62,13 +63,21 @@ describe.each(providers)("burn nfts", (ethereum: Ethereum) => {
 				uri: "ipfs://ipfs/hash",
 				royalties: [],
 			} as ERC721RequestV2)
+		if (minted.type === MintResponseTypeEnum.ON_CHAIN) {
+			await minted.transaction.wait()
+		}
 		const testBalance = await testErc721.functionCall("balanceOf", testAddress).call()
 		expect(toBn(testBalance).toString()).toBe("1")
 
-		await burn(checkChainId, {
-			contract: contractErc721,
-			tokenId: minted.tokenId,
+		const burnTx = await burn(checkChainId, {
+			assetType: {
+				contract: contractErc721,
+				tokenId: minted.tokenId,
+			},
 		})
+		if (burnTx) {
+			await burnTx.wait()
+		}
 		const testBalanceAfterBurn = await testErc721.functionCall("balanceOf", testAddress).call()
 		expect(toBn(testBalanceAfterBurn).toString()).toBe("0")
 	})
@@ -84,16 +93,25 @@ describe.each(providers)("burn nfts", (ethereum: Ethereum) => {
 				royalties: [],
 				supply: 100,
 			} as ERC1155RequestV1)
-		await burn(checkChainId, {
-			contract: contractErc1155,
-			tokenId: minted.tokenId,
-		}, toBigNumber("50"))
+		if (minted.type === MintResponseTypeEnum.ON_CHAIN) {
+			await minted.transaction.wait()
+		}
+		const burnTx = await burn(checkChainId, {
+			assetType: {
+				contract: contractErc1155,
+				tokenId: minted.tokenId,
+			},
+			amount: toBigNumber("50"),
+		})
+		if (burnTx) {
+			await burnTx.wait()
+		}
 
 		const testBalanceAfterBurn = await testErc1155.functionCall("balanceOf", testAddress, minted.tokenId).call()
 		expect(toBn(testBalanceAfterBurn).toString()).toBe("50")
 	})
 
-	test("should burn ERC-721 v3 lazy", async () => {
+	test.skip("should burn ERC-721 v3 lazy", async () => {
 		const minted = await mint(
 			mintLazyApi,
 			checkChainId,
@@ -105,12 +123,21 @@ describe.each(providers)("burn nfts", (ethereum: Ethereum) => {
 				lazy: true,
 			} as ERC721RequestV3)
 		await burn(checkChainId, {
-			contract: e2eErc721V3ContractAddress,
-			tokenId: minted.tokenId,
+			assetType: {
+				contract: e2eErc721V3ContractAddress,
+				tokenId: minted.tokenId,
+			},
+			creators: [{ account: toAddress(testAddress), value: 10000 }],
+		})
+		await retry(5, 2000, async () => {
+			const nftItemResponse = await apis.nftItem.getNftItemById({
+				itemId: `${e2eErc721V3ContractAddress}:${minted.tokenId}`,
+			})
+			expect(nftItemResponse.deleted).toBe(true)
 		})
 	})
 
-	test("should burn ERC-1155 v2 lazy", async () => {
+	test.skip("should burn ERC-1155 v2 lazy", async () => {
 		const minted = await mint(mintLazyApi, checkChainId, {
 			collection: createErc1155V2Collection(e2eErc1155V2ContractAddress),
 			uri: "ipfs://ipfs/hash",
@@ -120,8 +147,44 @@ describe.each(providers)("burn nfts", (ethereum: Ethereum) => {
 			lazy: true,
 		} as ERC1155RequestV2)
 		await burn(checkChainId, {
-			contract: e2eErc1155V2ContractAddress,
-			tokenId: minted.tokenId,
+			assetType: {
+				contract: e2eErc1155V2ContractAddress,
+				tokenId: minted.tokenId,
+			},
+			amount: toBigNumber("50"),
+			creators: [{ account: toAddress(testAddress), value: 10000 }],
+		})
+
+		await retry(5, 2000, async () => {
+			const nftItemResponse = await apis.nftItem.getNftItemById({
+				itemId: `${e2eErc1155V2ContractAddress}:${minted.tokenId}`,
+			})
+			expect(nftItemResponse.deleted).toBe(true)
+		})
+	})
+
+	test.skip("should burn ERC-1155 v2 lazy and burn creators is empty", async () => {
+		const minted = await mint(mintLazyApi, checkChainId, {
+			collection: createErc1155V2Collection(e2eErc1155V2ContractAddress),
+			uri: "ipfs://ipfs/hash",
+			supply: 100,
+			creators: [{ account: toAddress(testAddress), value: 10000 }],
+			royalties: [],
+			lazy: true,
+		} as ERC1155RequestV2)
+		await burn(checkChainId, {
+			assetType: {
+				contract: e2eErc1155V2ContractAddress,
+				tokenId: minted.tokenId,
+			},
+			creators: [],
+		})
+
+		await retry(5, 2000, async () => {
+			const nftItemResponse = await apis.nftItem.getNftItemById({
+				itemId: `${e2eErc1155V2ContractAddress}:${minted.tokenId}`,
+			})
+			expect(nftItemResponse.deleted).toBe(true)
 		})
 	})
 })