+++ packages/sdk/src/nft/deploy-erc721.test.ts
@@ -2,22 +2,24 @@ import { createE2eProvider } from "@rarible/ethereum-sdk-test-common"
 import Web3 from "web3"
 import { Web3Ethereum } from "@rarible/web3-ethereum"
 import { Configuration, GatewayControllerApi } from "@rarible/ethereum-api-client"
-import { send as sendTemplate } from "../common/send-transaction"
+import { getSendWithInjects } from "../common/send-transaction"
 import { getApiConfig } from "../config/api-config"
 import { getEthereumConfig } from "../config"
 import { e2eConfig } from "../config/e2e"
+import { checkChainId } from "../order/check-chain-id"
 import { DeployErc721 } from "./deploy-erc721"
 
 describe("deploy erc-721 token test", () => {
 	const { provider } = createE2eProvider()
 	const web3 = new Web3(provider)
-	const ethereum1 = new Web3Ethereum({ web3 })
+	const ethereum1 = new Web3Ethereum({ web3, gas: 2000000 })
 
 	const config = getEthereumConfig("e2e")
 	config.factories.erc721 = e2eConfig.factories.erc721
 	const configuration = new Configuration(getApiConfig("e2e"))
 	const gatewayApi = new GatewayControllerApi(configuration)
-	const send = sendTemplate.bind(null, gatewayApi)
+	const checkWalletChainId = checkChainId.bind(null, ethereum1, config)
+	const send = getSendWithInjects().bind(null, gatewayApi, checkWalletChainId)
 	const deployErc721 = new DeployErc721(ethereum1, send, config)